review: first batch of fix
