+++ .gitignore
@@ -13,3 +13,5 @@ book
 
 # Generated by the block traces example
 traces
+
+burntpix/svgs
