+++ bins/revm-test/Cargo.toml
@@ -9,6 +9,12 @@ bytes = "1.4"
 hex = "0.4"
 revm = { path = "../../crates/revm", version = "3.3.0",default-features=false }
 microbench = "0.5"
+alloy-sol-macro = "0.6.0"
+alloy-sol-types = "0.6.0"
+lazy_static = "1.4.0"
+regex = "1.10.3"
+eyre = "0.6.8"
+
 
 [[bin]]
 name = "analysis"
@@ -18,3 +24,6 @@ name = "snailtracer"
 
 [[bin]]
 name = "transfer"
+
+[[bin]]
+name = "burntpix"
