+++ bins/revm-test/src/bin/burntpix/main.rs
@@ -1,20 +1,20 @@
-use crate::burntpix::genesis_alloc::BURNTPIX_ADDRESS;
-use crate::burntpix::genesis_alloc::GENESIS_ALLOCS;
 use alloy_sol_macro::sol;
 use alloy_sol_types::SolCall;
-use criterion::Criterion;
+
+use genesis_alloc::{BURNTPIX_ADDRESS, GENESIS_ALLOCS};
 use regex::bytes::Regex;
 use revm::{
     db::{CacheDB, EmptyDB},
     primitives::{
-        address, hex, keccak256, AccountInfo, Bytecode, ExecutionResult, Output, TransactTo, U256,
+        address, hex, keccak256, AccountInfo, Bytecode, ExecutionResult, Output, TransactTo, B256,
+        U256,
     },
     Evm,
 };
-use revm_precompile::B256;
-use std::error::Error;
+
 use std::fs::File;
-use std::time::Duration;
+use std::{error::Error, time::Instant};
+
 use std::{io::Write, str::FromStr};
 pub mod genesis_alloc;
 
@@ -24,21 +24,16 @@ sol! {
         function run( uint32 seed, uint256 iterations) returns (string);
     }
 }
-const DEFAULT_SEED: &str = "0";
+
+const DEFAULT_SEED: &str = "0xF1FD58E";
 const DEFAULT_ITERATIONS: &str = "0x7A120";
 
-pub fn burntpix(c: &mut Criterion) {
+fn main() {
     let (seed, iterations) = try_init_env_vars().expect("Failed to parse env vars");
 
     let run_call_data = IBURNTPIX::runCall { seed, iterations }.abi_encode();
 
     let db = init_db();
-
-    let mut g = c.benchmark_group("burntpix");
-    g.noise_threshold(0.03)
-        .warm_up_time(Duration::from_secs(3))
-        .measurement_time(Duration::from_secs(130))
-        .sample_size(10);
 
     let mut evm = Evm::builder()
         .modify_tx_env(|tx| {
@@ -48,18 +43,14 @@ pub fn burntpix(c: &mut Criterion) {
         })
         .with_db(db)
         .build();
-
-    g.bench_function("burntpix".to_string(), |b| {
-        b.iter(|| evm.transact().unwrap())
-    });
-
-    // transact again to get the return data and create the svg
+    let started = Instant::now();
     let tx_result = evm.transact().unwrap().result;
     let return_data = match tx_result {
         ExecutionResult::Success {
             output, gas_used, ..
         } => {
             println!("Gas used: {:?}", gas_used);
+            println!("Time elapsed: {:?}", started.elapsed());
             match output {
                 Output::Call(value) => value,
                 _ => unreachable!("Unexpected output type"),
@@ -80,7 +71,7 @@ pub fn burntpix(c: &mut Criterion) {
 }
 fn svg(filename: String, svg_data: &[u8]) -> Result<(), Box<dyn Error>> {
     let current_dir = std::env::current_dir()?;
-    let svg_dir = current_dir.join("benches").join("burntpix").join("svgs");
+    let svg_dir = current_dir.join("burntpix").join("svgs");
     std::fs::create_dir_all(&svg_dir)?;
 
     let file_path = svg_dir.join(format!("{}.svg", filename));
@@ -92,10 +83,15 @@ fn svg(filename: String, svg_data: &[u8]) -> Result<(), Box<dyn Error>> {
 
 fn try_init_env_vars() -> Result<(u32, U256), Box<dyn Error>> {
     let seed_from_env = std::env::var("SEED").unwrap_or(DEFAULT_SEED.to_string());
-    let seed: u32 = seed_from_env.parse()?;
+    let seed: u32 = try_from_hex_to_u32(&seed_from_env)?;
     let iterations_from_env = std::env::var("ITERATIONS").unwrap_or(DEFAULT_ITERATIONS.to_string());
     let iterations = U256::from_str(&iterations_from_env)?;
     Ok((seed, iterations))
+}
+
+fn try_from_hex_to_u32(hex: &str) -> eyre::Result<u32> {
+    let trimmed = hex.strip_prefix("0x").unwrap_or(&hex);
+    u32::from_str_radix(trimmed, 16).map_err(|e| eyre::eyre!("Failed to parse hex: {}", e))
 }
 
 fn init_db() -> CacheDB<EmptyDB> {
