+++ crates/revm/Cargo.toml
@@ -39,10 +39,6 @@ ethers-contract = { version = "2.0.11", default-features = false }
 anyhow = "1.0.79"
 criterion = "0.5"
 indicatif = "0.17"
-alloy-sol-macro = "0.6.0"
-alloy-sol-types = "0.6.0"
-lazy_static = "1.4.0"
-regex = "1.10.3"
 
 [features]
 default = ["std", "c-kzg", "secp256k1"]
